# back-end/src/api/localizations/localization.dao.js
const CustomError = require('../../shared/custom-error');
const localizations = require('../../db/schemas/localizations.schema');

class LocalizationDAO {
    getAllLocalizations = async () => {
        const data = await localizations.find().catch(err => {
            console.log('error: ', err)
            throw new CustomError(err.message, 'MongoDB Error');
        });

        return data;
    }
}

module.exports = LocalizationDAO;