# back-end/src/swagger/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Customer Manager API",
    "description": "Documentation automatically generated by the <b>swagger.autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/api/v1/",
  "tags": [
    {
      "name": "KPI",
      "description": "Endpoints"
    }
  ],
  "schemes": [
    "http"
  ],
  "securityDefinitions": {},
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/customers/{cpf}/{uf}/{city}": {
      "get": {
        "tags": [
          "Customers"
        ],
        "description": "Get PF customer data from MongoDB database.",
        "parameters": [
          {
            "name": "cpf",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Customer document (CPF)."
          },
          {
            "name": "uf",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Customer UF."
          },
          {
            "name": "city",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Customer city."
          }
        ],
        "responses": {
          "200": {
            "description": "JSON data",
            "schema": {
              "$ref": "#/definitions/Customers"
            }
          },
          "400": {
            "description": "Validation Error",
            "schema": {
              "$ref": "#/definitions/CustomError"
            }
          },
          "500": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/CustomError"
            }
          }
        }
      }
    },
    "/api/v1/customers/{cnpj}/{uf}/{city}": {
      "get": {
        "tags": [
          "Customers"
        ],
        "description": "Get PJ customer data from MongoDB database.",
        "parameters": [
          {
            "name": "cnpj",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Customer document (CNPJ)."
          },
          {
            "name": "uf",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Customer UF."
          },
          {
            "name": "city",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Customer city."
          }
        ],
        "responses": {
          "200": {
            "description": "JSON data",
            "schema": {
              "$ref": "#/definitions/Customers"
            }
          },
          "400": {
            "description": "Validation Error",
            "schema": {
              "$ref": "#/definitions/CustomError"
            }
          },
          "500": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/CustomError"
            }
          }
        }
      }
    },
    "/api/v1/customers/": {
      "post": {
        "tags": [
          "Customers"
        ],
        "description": "Creates a new Customer if one does not already exist.",
        "parameters": [
          {
            "name": "type",
            "in": "body",
            "description": "Person type (PF/PJ).",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "body",
            "description": "Customer name.",
            "required": true,
            "type": "string"
          },
          {
            "name": "document",
            "in": "body",
            "description": "Customer document (CPF/CNPJ).",
            "required": true,
            "type": "string"
          },
          {
            "name": "phone",
            "in": "body",
            "description": "Customer phone number.",
            "required": true,
            "type": "string"
          },
          {
            "name": "uf",
            "in": "body",
            "description": "Customer UF.",
            "required": true,
            "type": "string"
          },
          {
            "name": "city",
            "in": "body",
            "description": "Customer city.",
            "required": true,
            "type": "string"
          },
          {
            "name": "birthDate",
            "in": "body",
            "description": "Customer birthDate.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Customers"
            },
            "description": "JSON data"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation Error",
            "schema": {
              "$ref": "#/definitions/CustomError"
            }
          },
          "500": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/CustomError"
            }
          }
        }
      }
    },
    "/api/v1/localizations/": {
      "get": {
        "tags": [
          "Localizations"
        ],
        "description": "Get all localizations from MongoDB database.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "JSON data",
            "schema": {
              "$ref": "#/definitions/Localizations"
            }
          },
          "400": {
            "description": "Validation Error",
            "schema": {
              "$ref": "#/definitions/CustomError"
            }
          },
          "500": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/CustomError"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Customers": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "6008d3cb14aeec5750cf8b0e"
        },
        "type": {
          "type": "string",
          "example": "PF"
        },
        "name": {
          "type": "string",
          "example": "Denis Slapelis"
        },
        "document": {
          "type": "number",
          "example": 46989096861
        },
        "phone": {
          "type": "string",
          "example": "(11) 98964-3889"
        },
        "uf": {
          "type": "string",
          "example": "SP"
        },
        "city": {
          "type": "string",
          "example": "São Paulo"
        },
        "birthDate": {
          "type": "string",
          "example": "1996-12-10T02:00:00.000Z"
        },
        "__v": {
          "type": "number",
          "example": 0
        }
      },
      "xml": {
        "name": "Customers"
      }
    },
    "Localizations": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "6008aec4ccb25a49a465e457"
          },
          "UF": {
            "type": "string",
            "example": "RO"
          },
          "city": {
            "type": "string",
            "example": "Ji-Paraná"
          },
          "__v": {
            "type": "number",
            "example": 0
          }
        }
      },
      "xml": {
        "name": "Localizations"
      }
    },
    "CustomError": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Error type"
        },
        "message": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "stack": {
          "type": "string",
          "example": "If env var SHOW_QUERY is true, the stacktrace of the error will be displayed, otherwise it will be null"
        }
      },
      "xml": {
        "name": "CustomError"
      }
    }
  }
}